requirements.txt
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cloudinary==1.22.0
cryptography==3.2.1
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.1.2
django-allauth==0.43.0
django-bootstrap3==14.2.0
django-cloudinary-storage==0.3.0
django-filter==2.4.0
django-heroku==0.3.1
django-tinymce==3.1.0
djangorestframework==3.12.1
gunicorn==20.0.4
idna==2.10
oauthlib==3.1.0
Pillow==8.0.1
psycopg2==2.8.6
pycparser==2.20
PyJWT==1.7.1
python-decouple==3.3
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.25.11
whitenoise==5.2.0
